requirements.txt
Django==5.2.4
requests==2.32.3
pandas==2.0.3
openpyxl==3.1.2
torch==2.1.0
transformers==4.35.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
easyocr==1.7.0
torchvision==0.16.0
accelerate==0.24.0
